requirements.txt
python-twitter
pymysql
sqlalchemy
sqlalchemy_utils
pyyaml